requirements.txt
numpy>=1.21.0
matplotlib==3.7.1
pandas>=2.1.3
seaborn>=0.12.2
scikit-learn>=1.0.0
scipy>=1.10.0
statsmodels>=0.14.0
pyyaml>=6.0
